+++ Sources/Ignite/Styles/SpacingType.swift
@@ -8,7 +8,7 @@
 /// A type that represents spacing values in either exact pixels or semantic spacing amounts.
 enum SpacingType {
     /// An exact spacing value in pixels.
-    case exact(Double)
+    case exact(Int)
 
     /// A semantic spacing value that adapts based on context.
     case semantic(SpacingAmount)
